Make filesystem as `intptr_t`

+++ src/xrCore/memory/xrMemory.h
@@ -42,7 +42,7 @@ extern XRCORE_API xrMemory Memory;
 
 // generic "C"-like allocations/deallocations
 template <class T>
-IC T*		xr_alloc	(u32 count)				{	return  (T*)Memory.mem_alloc(count*sizeof(T));	}
+IC T*		xr_alloc	(size_t count)			{	return  (T*)Memory.mem_alloc(count*sizeof(T));	}
 template <class T>
 IC void		xr_free		(T* &P)					{	if (P) { Memory.mem_free((void*)P); P=NULL;	};	}
 IC void*	xr_malloc	(size_t size)			{	return	Memory.mem_alloc(size);					}
