+++ src/xrEngine/IGame_Level.h
@@ -95,7 +95,7 @@ public:
 	virtual BOOL				Load					( u32 dwNum );
 	virtual BOOL				Load_GameSpecific_Before( )										{ return TRUE; };		// before object loading
 	virtual BOOL				Load_GameSpecific_After	( )										{ return TRUE; };		// after object loading
-	virtual void				Load_GameSpecific_CFORM	( CDB::TRI* T, u32 count )				= 0;
+	virtual void				Load_GameSpecific_CFORM	( CDB::TRI* T, size_t count )			= 0;
 
 	virtual void	_BCL		OnFrame					( void );
 	virtual void				OnRender				( void );
